Updated format
Settled on the formatting. for the career stuff. Need to decide on whether coding project goes in each annual card or its own card. Need to fix formatting on the career stuff to the right of the timeline. Not even currently.

Maybe the spacing needs to be a percentage related to the height of the card.